+++ src/de/dhbw/Microcontroller/Befehle/PIC/BCF.java
@@ -16,11 +16,11 @@ public class BCF extends Instruction {
 
     @Override
     public void execute(){
-        byte b = (byte) argument1;
+        int b = argument1;
         int f = argument2;
 
         // TODO: Hier stimmt was nicht
-        b = (byte) (b & ~(1 << (f-1))); //das bit f wird in b  auf 0 gesetzt
+        b = (b & ~(1 << (f-1))); //das bit f wird in b  auf 0 gesetzt
         memory.setAddress(f, b);
         incrementProgramCounter();
