+++ src/de/dhbw/Microcontroller/Befehle/PIC/BSF.java
@@ -15,10 +15,10 @@ public class BSF extends Instruction {
 
     @Override
     public void execute(){
-        byte b = (byte) argument1;
+        int b = argument1;
         int f = argument2;
 
-        b = (byte) (b |(1 << (f-1))); //das bit f wird in b  auf 1 gesetzt
+        b = (b |(1 << (f-1))); //das bit f wird in b  auf 1 gesetzt
 
         incrementProgramCounter();
     }
