+++ src/de/dhbw/Microcontroller/Befehle/PIC/MOVF.java
@@ -18,8 +18,8 @@ public class MOVF extends Instruction {
         byte d = (byte) argument1;
         byte f = (byte) argument2;
 
-        Byte fValue  = memory.getAddress(f);
-        fValue = (byte) (~ fValue & 0xFF);
+        int fValue  = memory.getAddress(f);
+        fValue = (~ fValue & 0xFF);
 
         if (d == 0)
             memory.setRegisterW(fValue);
