+++ src/de/dhbw/Model/MemoryView.java
@@ -5,77 +5,77 @@ package de.dhbw.Model;
  */
 public class MemoryView {
     private String memoryRow;
-    private Byte column0;
-    private Byte column1;
-    private Byte column2;
-    private Byte column3;
-    private Byte column4;
-    private Byte column5;
-    private Byte column6;
-    private Byte column7;
+    private int column0;
+    private int column1;
+    private int column2;
+    private int column3;
+    private int column4;
+    private int column5;
+    private int column6;
+    private int column7;
 
 
-    public Byte getColumn0() {
+    public int getColumn0() {
         return column0;
     }
 
-    public void setColumn0(Byte column0) {
+    public void setColumn0(int column0) {
         this.column0 = column0;
     }
 
-    public Byte getColumn1() {
+    public int getColumn1() {
         return column1;
     }
 
-    public void setColumn1(Byte column1) {
+    public void setColumn1(int column1) {
         this.column1 = column1;
     }
 
-    public Byte getColumn2() {
+    public int getColumn2() {
         return column2;
     }
 
-    public void setColumn2(Byte column2) {
+    public void setColumn2(int column2) {
         this.column2 = column2;
     }
 
-    public Byte getColumn3() {
+    public int getColumn3() {
         return column3;
     }
 
-    public void setColumn3(Byte column3) {
+    public void setColumn3(int column3) {
         this.column3 = column3;
     }
 
-    public Byte getColumn4() {
+    public int getColumn4() {
         return column4;
     }
 
-    public void setColumn4(Byte column4) {
+    public void setColumn4(int column4) {
         this.column4 = column4;
     }
 
-    public Byte getColumn5() {
+    public int getColumn5() {
         return column5;
     }
 
-    public void setColumn5(Byte column5) {
+    public void setColumn5(int column5) {
         this.column5 = column5;
     }
 
-    public Byte getColumn6() {
+    public int getColumn6() {
         return column6;
     }
 
-    public void setColumn6(Byte column6) {
+    public void setColumn6(int column6) {
         this.column6 = column6;
     }
 
-    public Byte getColumn7() {
+    public int getColumn7() {
         return column7;
     }
 
-    public void setColumn7(Byte column7) {
+    public void setColumn7(int column7) {
         this.column7 = column7;
     }
